Password reset, Shortener

+++ src/Controller/PortfolioController.php
@@ -9,11 +9,13 @@ use src\Exception\AppException;
 use src\Exception\ConfigurationException;
 use src\Exception\NotFoundException;
 use src\Exception\StorageException;
-use src\Model\Auth;
+use src\Model\Link;
+use src\Model\User;
 use src\Model\Weather;
 use src\Utils\Request;
 use src\Utils\Response;
 use src\View;
+use Google\GTrends;
 
 class PortfolioController
 {
@@ -37,12 +39,16 @@ class PortfolioController
             'chat' => "chat",
             'geo' => "geo",
             'trends' => "trends",
+            'short' => "short",
+            'deleteShort' => "deleteShort",
             'error' => "error",
-            'authLogin' => "authLogin",
-            'authRegister' => "authRegister",
-            'authUpdate' => "authUpdate",
-            'authDelete' => "authDelete",
-            'authLogout' => "authLogout",
+            'authLogin' => "auth/login",
+            'authRegister' => "auth/register",
+            'authUpdate' => "auth/update",
+            'authDelete' => "auth/delete",
+            'authLogout' => "auth/logout",
+            'authReset' => "auth/reset",
+            'authForget' => "auth/forget",
         };
 
         if($params->isPost()){
@@ -61,7 +67,55 @@ class PortfolioController
      * @throws NotFoundException
      */
     private function site(string $page, Request $request){
-        $this->view->render($page, $request->getData());
+        switch ($page){
+            case 'short':
+                if(!empty($request->getParam('id'))){
+                    $db = new Link(require_once("config/config.php"));
+                    $long = ($db->readOne($request->getParam('id')))[0]['long'];
+                    header("Location: $long");
+                }
+                elseif(empty($request->sessionParam('user')['id'])){
+                    header("Location: /Login");
+                }
+                else{
+                    $db = new Link(require_once("config/config.php"));
+                    $this->view->render($page, $request->getData(), ['links' => $db->read($request->sessionParam('user')['id'])]);
+                }
+                break;
+            case 'geo':
+                if(!empty($request->getParam("ip"))){
+                    $geo = json_decode(file_get_contents("http://ip-api.com/json/{$_GET["ip"]}?fields=status,country,city,lat,lon,query"));
+                }
+                else{
+                    $geo = json_decode(file_get_contents("http://ip-api.com/json/{$_SERVER['REMOTE_ADDR']}?fields=status,country,city,lat,lon,query"));
+                }
+                $this->view->render($page, $request->getData(), ['geo' => $geo]);
+                break;
+            case 'trends':
+
+                /**
+                 * @throws AppException
+                 */
+
+                $options = [
+                    'hl' => 'en-US',
+                    'tz' => 360,
+                    'geo' => 'US',
+                ];
+                try {
+                    $gt = new GTrends($options);
+                    $this->view->render($page, $request->getData(), ['trends' => $gt->getDailySearchTrends()['default']['trendingSearchesDays']]);
+                }
+                catch (\Exception $e) {
+                    throw new AppException("There was an error while trying to show trends " . $e, 408);
+                }
+                break;
+            default:
+                $this->view->render($page, $request->getData());
+                break;
+
+        }
+
     }
 
     /**
@@ -82,28 +136,50 @@ class PortfolioController
                     $db = new Weather(require_once("config/config.php"));
                     new Response($db->read($request));
                     return;
-                case 'authLogin':
-                    $db = new Auth(require_once("config/config.php"));
-                    new Response($db->read($request));
-                    return;
-                case 'authRegister':
-                    $db = new Auth(require_once("config/config.php"));
+                case 'short':
+                    $db = new Link(require_once("config/config.php"));
                     new Response($db->create($request));
                     return;
-                case 'authUpdate':
-                    $db = new Auth(require_once("config/config.php"));
-                    new Response($db->update($request));
+                case 'deleteShort':
+                    $auth = new AuthController(new User(require("config/config.php")));
+                    if ($auth->auth($request)){
+                        $db = new Link(require("config/config.php"));
+                        new Response($db->delete($request));
+                    }
+                    else{
+                        new Response(['error' => 'auth']);
+                    }
                     return;
-                case 'authDelete':
-                    $db = new Auth(require_once("config/config.php"));
-                    new Response($db->delete($request));
+                case 'auth/login':
+                    $auth = new AuthController(new User(require_once("config/config.php")));
+                    new Response($auth->login($request));
+                    return;
+                case 'auth/register':
+                    $auth = new AuthController(new User(require_once("config/config.php")));
+                    new Response($auth->add($request));
+                    return;
+                case 'auth/update':
+                    $auth = new AuthController(new User(require_once("config/config.php")));
+                    new Response($auth->edit($request));
+                    return;
+                case 'auth/delete':
+                    $auth = new AuthController(new User(require_once("config/config.php")));
+                    new Response($auth->remove($request));
+                    return;
+                case 'auth/forget':
+                    $auth = new AuthController(new User(require_once("config/config.php")));
+                    new Response($auth->forget($request));
+                    return;
+                case 'auth/reset':
+                    $auth = new AuthController(new User(require_once("config/config.php")));
+                    new Response($auth->reset($request));
                     return;
             }
 
         }
         catch(PDOException $e)
         {
-            throw new AppException('There was an error while trying to get data',500);
+            throw new AppException('There was an error while trying to get data ' . $e,500);
         }
     }
 }
